+++ tests/testthat/test-hp.R
@@ -60,10 +60,10 @@ test_that("hp fitdists with no dists", {
 })
 
 test_that("hp fitdists", {
-  expect_equal(ssd_hp(boron_dists, 1)$est, 3.66480402177009)
+  expect_equal(ssd_hp(boron_dists, 1)$est, 3.6651119524728)
   expect_equal(
     ssd_hp(boron_dists, c(0, 1, 30, Inf))$est,
-    c(0, 3.66480402177009, 72.8682760255365, 100)
+    c(0, 3.6651119524728, 72.8664372774006, 100)
   )
 })
 
@@ -77,6 +77,6 @@ test_that("hp fitdistcens", {
 test_that("hp fitdistscens", {
   expect_equal(
     ssd_hp(fluazinam_dists, c(0, 1, 30, Inf))$est,
-    c(0, 4.2330766654294, 26.9580291772613, 100)
+    c(0, 4.23518701256203, 26.9619242506997, 100)
   )
 })