+++ tests/testthat/test-llog.R
@@ -14,11 +14,9 @@
 
 context("llog")
 
-test_that("rllog", {
-  set.seed(99)
-  r <- rllog(100000)
-  expect_identical(length(r), 100000L)
-  expect_equal(mean(log(r)), 1, tolerance = 0.1)
+test_that("dllog", {
+  expect_equal(dllog(exp(3), log(3), log(1)), 0.003720046, tolerance = 0.0000001)
+  expect_equal(dllog(exp(4), log(3), log(1)), 0.001200358, tolerance = 0.0000001)
 })
 
 test_that("fit llog", {
@@ -26,17 +24,19 @@ test_that("fit llog", {
 
   expect_true(is.fitdist(dist))
   expect_equal(coef(dist),
-  c(scale = 2.6261248978507, shape = 0.740309228071107))
+  c(scalelog = 0.965466010495141, shapelog = -0.300741556664549))
 })
 
 test_that("pqllog", {
-  expect_equal(log(qllog(0.5)), 1, tolerance = 0.000001)
-  expect_equal(pllog(exp(3)), 0.8807971, tolerance = 0.0000001)
-  expect_equal(pllog(exp(4)), 0.9525741, tolerance = 0.0000001)
+  expect_equal(log(qllog(0.5, 0, 0)), 1, tolerance = 0.000001)
+  expect_equal(pllog(exp(3), 0, 0), 0.8807971, tolerance = 0.0000001)
+  expect_equal(pllog(exp(4), 0, 0), 0.9525741, tolerance = 0.0000001)
   expect_identical(pllog(qllog(0.5, 3, 1), 3, 1), 0.5)
 })
 
-test_that("dllog", {
-  expect_equal(dllog(exp(3), 3, 1), 0.003720046, tolerance = 0.0000001)
-  expect_equal(dllog(exp(4), 3, 1), 0.001200358, tolerance = 0.0000001)
+test_that("rllog", {
+  set.seed(99)
+  r <- rllog(100000, 0, 0)
+  expect_identical(length(r), 100000L)
+  expect_equal(mean(log(r)), 1, tolerance = 0.1)
 })